Extract chess evaluation in ChessEvaluation.js file.
